requirements.txt
Pyrogram
TgCrypto
python-dotenv